Removed geomorph dependency - version 0.1.1
As far as I can tell, the geomorph function procD.pgls cannot be called
from within a wrapper function because it looks for global variables
corresponding to the argument f1. As a (hopefully) temporary fix, I have
slightly modified procD.pgls as an internal function in phylocurve
called procD.pgls2 (which means it was also necessary to copy over a
bunch of other geomorph internal functions into phyocurve).